22th commit

+++ Web/Models/Order.cs
@@ -9,7 +9,7 @@ namespace laptop_store.Models
     {
         public int OrderID { get; set; }
         public string UserEmail { get; set; }
-        public float OrderPrice { get; set; }
+        public double OrderPrice { get; set; }
         public DateTime OrderDate { get; set; }
     }
 }
